pyproject.toml
[tool.poetry]
name = "resume"
version = "0.1.0"
description = "Christopher Dewan's Resume"
authors = ["Chris Dewan <chris.dewan@m3rlin.net>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
click = "^8.0.3"
Jinja2 = "^3.0.2"
pyspellchecker = "^0.6.2"
pytest = "^6.2.5"
PyLaTeX = "^1.4.1"
validators = "^0.18.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
